electron/assets/python/requirements.txt
torch>=2.2.2
torchaudio>=2.2.2
torchvision>=0.17.2
transformers>=4.48.1
sentence-transformers>=3.4.1
datasets>=3.2.0
geopy>=2.4.1
openai>=1.59.3
spacy>=3.8.4
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5
whisper>=1.0.0
tokenizers>=0.15.2
sqlparse>=0.5.3
spacy-alignments>=0.9.1
spacy-lookups-data>=1.0.5
spacy-transformers>=1.3.5
ffmpeg-python>=0.2.0
contractions>=0.1.73
auto-labeling-pipeline>=0.1.23
python-dotenv>=1.0.1 